Checkpoint 5: Added user roles and authorization for Wiki CRUD.

+++ app/controllers/wikis_controller.rb
@@ -5,6 +5,7 @@ class WikisController < ApplicationController
 
   def new
      @wiki = Wiki.new
+     authorize @wiki
   end
 
   def show
@@ -13,10 +14,12 @@ class WikisController < ApplicationController
 
   def edit
      @wiki = Wiki.find(params[:id])
+     authorize @wiki
   end
 
   def create
      @wiki = Wiki.new
+     authorize @wiki
      @wiki.title = params[:wiki][:title]
      @wiki.body = params[:wiki][:body]
      @wiki.user = current_user
@@ -33,19 +36,21 @@ class WikisController < ApplicationController
 
   def update
      @wiki = Wiki.find(params[:id])
+     authorize @wiki
      @wiki.assign_attributes(params.require(:wiki).permit(:title, :body))
 
      if @wiki.save
         flash[:notice] = "Wiki updated."
         redirect_to [@wiki]
      else
-        flash.now[:alert] = "There was an error when trying to save the wiki. PLease try again."
+        flash.now[:alert] = "There was an error when trying to save the wiki. Please try again."
         render :edit
      end
   end
 
   def destroy
      @wiki = Wiki.find(params[:id])
+     authorize @wiki
 
      if @wiki.destroy
         flash[:notice] = "Wiki deleted."
